refactor: 딜러 게임 결과가 승 무 패 순으로 출력되게 변경
